backend/requirements.txt
asgiref==3.8.1
Django==4.2  # Use this stable version
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
channels==4.1.0
django-allauth
djangorestframework-simplejwt
requests